+++ src/pages/Dashboard/Dashboard.jsx
@@ -1,5 +1,6 @@
 import { Navigate, Route, Routes, useNavigate, useLocation } from 'react-router-dom'
 import { Student, Teacher } from '../index'
+import Create from '../Teacher/Create/Create'
 import { useEffect } from 'react'
 import { useSelector } from 'react-redux'
 
@@ -10,17 +11,18 @@ function Dashboard() {
     const userRole = useSelector((state) => state.user.userRole)
 
     useEffect(() => {
-        if(location.pathname !== `/dashboard/${userRole}`){
+        if(location.pathname !== `/dashboard/${userRole}` && location.pathname !== `/dashboard/${userRole}/create`){
             navigate('/')
         }
     }, [location])
-
+    
     return (
         <>
             <Routes>
                 <Route path='/' element={<Navigate to={'/'} />} />
                 <Route path='/student' element={<Student />} />
                 <Route path='/teacher' element={<Teacher />} />
+                <Route path='/teacher/create' element={<Create />} />
             </Routes>
         </>
     )
