+++ src/pages/Teacher/ActiveCourses/ActiveCourses.module.css
@@ -1,0 +1,45 @@
+.coursescontainer{
+    display: flex;
+    flex-direction: row;
+    justify-content: space-evenly;
+}
+
+
+
+.courses{
+    display: grid;
+    grid-template-rows: repeat(3, minmax(0, 1fr));
+    grid-template-columns: repeat(3, minmax(0, 1fr));
+    justify-content: center;
+    align-items: center;
+}
+
+.coursestitle{
+    text-align: center;
+}
+
+.heard {
+    position: relative;
+    width: 100vw;
+    height: 652px;
+    z-index: 15;
+    margin-bottom: 2rem;
+}
+
+.heard img {
+    object-fit: cover;
+    width: 100%;
+    height: 100%
+}
+
+.filters{
+    display: flex;
+    flex-direction: column;
+    width: 16rem;
+    border-right: 2px solid black;
+    padding-right: 2vw;
+}
+
+.filters button{
+    font-size: 1.35rem;
+}
