+++ src/pages/index.js
@@ -1,10 +1,10 @@
 export { default as About } from './About/About';
 export { default as Courses } from './Courses/Courses';
-export { default as Create } from './Create/Create';
 export { default as Dashboard } from './Dashboard/Dashboard';
 export { default as Detail } from './Detail/Detail';
 export { default as Donate } from './Donate/Donate';
 export { default as Gracias } from './Gracias/Gracias';
 export { default as Student } from './Student/Student';
+export { default as Create } from './Teacher/Create/Create';
 export { default as Teacher } from './Teacher/Teacher';
 export { default as Home } from './home/Home';
